Fix compiler warnings.

+++ src/main/java/net/java/otr4j/io/messages/DataMessage.java
@@ -35,12 +35,10 @@ public final class DataMessage extends AbstractEncodedMessage {
     public final byte[] mac;
     public final byte[] oldMACKeys;
 
-    public DataMessage(@Nonnull final MysteriousT t, @Nonnull final byte[] mac,
-            @Nonnull final byte[] oldMacKeys, final int senderInstanceTag,
-            final int receiverInstanceTag) {
-        this(t.protocolVersion, (byte) t.flags, t.senderKeyID, t.recipientKeyID,
-                t.nextDH, t.ctr, t.encryptedMessage, mac, oldMacKeys,
-                senderInstanceTag, receiverInstanceTag);
+    public DataMessage(@Nonnull final MysteriousT t, @Nonnull final byte[] mac, @Nonnull final byte[] oldMacKeys,
+                       final int senderInstanceTag, final int receiverInstanceTag) {
+        this(t.protocolVersion, t.flags, t.senderKeyID, t.recipientKeyID, t.nextDH, t.ctr, t.encryptedMessage, mac,
+            oldMacKeys, senderInstanceTag, receiverInstanceTag);
     }
 
     public DataMessage(final int protocolVersion, final byte flags, final int senderKeyID,
